refactor element node parsers into base class
